@@ -2,6 +2,7 @@ import Vue from 'vue';
 import App from './components/App.vue';
 
 new Vue({
-  el: '#app',
+  el    : '#app',
   render: h => h(App),
 });
+
